Fix the transformed time when the iteration duration is infinity
A group should pass a non-zero transformed time to its children whenever its
directed time is non-zero, even if its iteration duration is infinite.
Currently, however, such a TimedItem always has a zero transformed time,
because this is calculated from the iteration fraction, which is always zero
when the iteration duration is infinity.

This change fixes the problem by bypassing the iteration fraction in this case
and setting the transformed time to the directed time.