Five small typos
1.2 1st paragraph
 all provide mechanisms \ that generate animated content
->
 all provide mechanisms that generate animated content
(remove backslash)

1.2 4th paragraph
and their execution order to defined such that
->
and their execution order is defined such that
(to becomes is)

3.5.15.1
The procedure to silently update the player playback rate of player, player tonew playback rate is identical to
->
The procedure to silently update the player playback rate of player, player to new playback rate is identical to
(add space)

3.10.1
Following this further transformations
->
Following this, further transformations
(add comma)

5.3
The playback rate of this player. Setting this attribute follows the procedure toupdate
->
The playback rate of this player. Setting this attribute follows the procedure to update
(add space)