Make setting the current time complete a pending pause, not abort it
The point of making pausing async is to allow time to sync up the
current time with the compositor. Setting the current time manually
should simply force it to the specified time and complete the pause
action, not abort it. (We do a similar thing for a pending play. For a
pending play we're waiting to establish a suitable start time. Manually
setting the start time in that case simply forces the start time to the
specified time and completes the play operation.)