Resolve start time on finish()
We have already resolved to make calling finish() clear the pause state
(see
https://lists.w3.org/Archives/Public/public-fx/2015AprJun/0038.html,
item 2). It turns out we need to do this when the animation is
play-pending as well otherwise if we simply cancel the pending play task
we'll end up in an odd state where the start time never gets resolved
and hence the animation never enters the finished state.

To address both these issues we need to fill in an unresolved start time
when finishing an animation. The one exception is when we don't have an
active timeline. Without an active timeline we can't enter the running
state but instead we should stay pending. Since "finished" is ultimately
a subset of running it makes sense that we can't finish the animation in
this case either but instead should remain pending.