Fix update the finished state procedure
We shouldn't create a new finished promise each time we become newly
un-finished. Otherwise, if we have a series of transitions from, e.g.
playing/finished/playing/finished/playing/finished and we register
listeners on the finished promise along the way, we'll end up with
several unique finished promises all of which should be resolved
asynchronously.